% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli.R
\name{bq_cli}
\alias{bq_cli}
\title{Run the bubblequiz command-line interface}
\usage{
bq_cli(args = commandArgs(trailingOnly = TRUE))
}
\arguments{
\item{args}{Command-line arguments, defaulting to those of the current session.}
}
\value{
Invisibly, the result of the dispatched command.
}
\description{
Run the bubblequiz command-line interface
}
